Added a GPL-v2 LICENSE.
According to http://thbecker.net/free_software_utilities/emacs_lisp/emacros/start_page.html
the original code was created in Mach 2007, that's before GPL-v3 was created.
The emacros.el has text that is a GPL license, and that seems to be GPL-v2.
I sent an email to Thomas Becker (the original author) but 'm still waiting for a reply.